+++ factory/src/test/java/io/github/factoryfx/factory/attribute/dependency/FactoryPolymorphicListAttributeTest.java
@@ -1,14 +1,15 @@
 package io.github.factoryfx.factory.attribute.dependency;
 
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
 import io.github.factoryfx.factory.FactoryTreeBuilderBasedAttributeSetup;
+import io.github.factoryfx.factory.SimpleFactoryBase;
 import io.github.factoryfx.factory.builder.FactoryTreeBuilder;
 import io.github.factoryfx.factory.jackson.ObjectMapperBuilder;
-import io.github.factoryfx.factory.SimpleFactoryBase;
 import io.github.factoryfx.factory.metadata.FactoryMetadataManager;
 import io.github.factoryfx.factory.testfactories.poly.ErrorPrinterFactory;
 import io.github.factoryfx.factory.testfactories.poly.Printer;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
 
 public class FactoryPolymorphicListAttributeTest {
 
@@ -38,7 +39,7 @@ public class FactoryPolymorphicListAttributeTest {
     public void test_newValue_noexception(){
         FactoryTreeBuilder<Object,PolymorphicFactoryExample> builder = new FactoryTreeBuilder<>(PolymorphicFactoryExample.class, ctx -> new PolymorphicFactoryExample());
         PolymorphicFactoryExample root = builder.buildTreeUnvalidated();
-        root.internal().serFactoryTreeBuilderBasedAttributeSetupForRoot(new FactoryTreeBuilderBasedAttributeSetup<>(builder));
+        root.internal().setFactoryTreeBuilderBasedAttributeSetupForRoot(new FactoryTreeBuilderBasedAttributeSetup<>(builder));
         root.referenceList.internal_createNewPossibleValues(FactoryMetadataManager.getMetadata(PolymorphicFactoryExample.class).getAttributeMetadata(f->f.referenceList));
     }